Working? Sleeve detector (using telemetry.)
Use Robot.initOD in INIT state of opmode, then use Robot.detectLevel in the RUN state of opmode. MAKE SURE to upload the model to the robot before running. Can be done with the Manage section of the robot's website (didn't work last year i think) OR manually uploading it to the phone AND changing the path to be absolute in the code (Robot.TFOD_MODEL_FILE). Note: may move tfod/vuforia final constants over to RobotHardware or RobotInfo later.