copied working PID for robobobo
made a comment about the logic for robobobo in javadoc on helper method and comment above the logic. haven't tested on final yet, but to find the negations on final, just test it and see which motors need reversing.
TL;DR works on robobobo, logic in comment, test final to see negation logic for that.